+++ assemblyscript/@gl/api/types/sound.ts
@@ -1,4 +1,13 @@
-export class SoundOpts {
+class Sprite {
+  /** The name of the sprite. */
+  name!: string;
+  /** The start time of the sprite in seconds. */
+  start: f32 = 0;
+  /** The length of the sprite in seconds. */
+  length: f32 = 0;
+}
+
+export class LoadOpts {
   /** The name of the sound to load from your `sounds` folder */
   name!: string;
   /** Whether the sound should autoplay on loads. Good for music. */
@@ -7,6 +16,14 @@ export class SoundOpts {
   loop: boolean = false;
   /** The volume of the sound. */
   volume: f32 = 1;
+  sprites: Sprite[] = [];
+}
+
+export class PlayOpts {
+  /** The ID of the sound to play. */
+  id!: i32;
+  /** The name of the sound to load from your `sounds` folder */
+  spriteId: string = "";
 }
 
 export class LoadAndPlaySoundOpts {