+++ scripts/reset.py
@@ -9,28 +9,43 @@ TILED_DIR = LEVEL_DIR / "tiled"
 ART_DIR = LEVEL_DIR / "art"
 SOUNDS_DIR = LEVEL_DIR / "sounds"
 
-PRESERVE_TILED = ["blank.tmj", "level.tiled-project", "level.tiled-session"]
+PRESERVE_TILED = [
+    "blank.tmj",
+    "blank.tiled-session",
+    "level.tiled-project",
+]
 
 
 def clear_tiled():
     for file in TILED_DIR.iterdir():
         if file.name not in PRESERVE_TILED:
             file.unlink()
-    blank = TILED_DIR / "blank.tmj"
-    if blank.exists():
-        shutil.copy(blank, TILED_DIR / "level.tmj")
+    blank_map = TILED_DIR / "blank.tmj"
+    if blank_map.exists():
+        shutil.copy(blank_map, TILED_DIR / "level.tmj")
+
+    blank_session = TILED_DIR / "blank.tiled-session"
+    if blank_session.exists():
+        shutil.copy(blank_session, TILED_DIR / "level.tiled-session")
 
 
 def clear_art():
     if ART_DIR.exists():
         shutil.rmtree(ART_DIR)
     ART_DIR.mkdir()
+    (ART_DIR / "restricted").mkdir()
 
 
 def clear_sounds():
     if SOUNDS_DIR.exists():
         shutil.rmtree(SOUNDS_DIR)
     SOUNDS_DIR.mkdir()
+    (SOUNDS_DIR / "restricted").mkdir()
+
+
+def clear_dialogue():
+    # FIXME
+    pass
 
 
 def main():
@@ -41,6 +56,7 @@ def main():
     clear_tiled()
     clear_art()
     clear_sounds()
+    clear_dialogue()
 
 
 if __name__ == "__main__":